% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlackScholes.R
\name{exactinformationmatrix}
\alias{exactinformationmatrix}
\title{Compute the exact information matrix of GBM}
\usage{
exactinformationmatrix(x, del, param)
}
\arguments{
\item{x}{Observation of the state variable at time t}

\item{del}{The time step between the current and previous observation}

\item{param}{The parameter 2-vector (mu,sigma)}
}
\description{
Compute the exact information matrix of GBM
}
\examples{
exactinformationmatrix(c(0.1,0.2,0.3,0.4,0.5,0.6),0.1,c(0.1,0.2))

}
