% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChain.R
\name{getChain}
\alias{getChain}
\title{Load the option chain from a csv file into a dataframe}
\usage{
getChain(fileName)
}
\arguments{
\item{fileName}{Filename of the csv file storing}
}
\description{
Load the option chain from a csv file into a dataframe
}
