% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdensity2info.R
\name{logdensity2info}
\alias{logdensity2info}
\title{Compute the information matrix = Sum_{i=1}^n Score_i' Score_i}
\usage{
logdensity2info(logdensity, x, del, param, args = NULL)
}
\arguments{
\item{logdensity:}{the model log likelihood function}

\item{x:}{Time series of the observed state variables}

\item{del:}{The uniform time step between observations}

\item{param:}{The parameter vector}
}
\description{
Compute the information matrix = Sum_{i=1}^n Score_i' Score_i
}
\examples{
 logdensity2info(ModelU1,x,0.1,c(0.01,0.2))

}
