% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdensity2loglik.R
\name{logdensity2loglik}
\alias{logdensity2loglik}
\title{Compute the log likelihood function of the diffusion model}
\usage{
logdensity2loglik(logdensity, x, del, param, args = NULL)
}
\arguments{
\item{logdensity:}{the model log likelihood function}

\item{x:}{Time series of the observed state variables}

\item{del:}{The uniform time step between observations}

\item{param:}{The parameter vector}
}
\description{
Compute the log likelihood function of the diffusion model
}
\examples{
 logdensity2loglik(ModelU1,x,0.1,c(0.01,0.2))

}
