% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelB1.R
\name{ModelB1}
\alias{ModelB1}
\title{Compute the maximum likelihood estimate of Model B1}
\usage{
ModelB1(x, x0, del, param)
}
\arguments{
\item{x}{Observation of the state variable at time t}

\item{x0}{Observation of the state variable at time t-1}

\item{del}{The time step between the current and previous observation}

\item{param}{The parameter 6-vector (r,h,a,b,c,d)}
}
\description{
Compute the maximum likelihood estimate of Model B1
}
\examples{
ModelB1(0.4,0.3,0.1,c(0.1,0.2,0.3,0.4,0.5,0.6))

}
