% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelB2.R
\name{ModelB2}
\alias{ModelB2}
\title{Compute the maximum likelihood estimate of Model B2}
\usage{
ModelB2(x, x0, del, param)
}
\arguments{
\item{x:}{Observation of the state variable at time t}

\item{x0:}{Observation of the state variable at time t-1}

\item{del:}{The time step between the current and previous observation}

\item{param:}{The parameter 12-vector (a0,a1,a2,b0,b1,b2,c0,c1,c2,d0,d1,d2)}
}
\description{
Compute the maximum likelihood estimate of Model B2
}
\examples{
ModelB2(0.4,0.3,0.1,c(0.1,0.2,0.3,0.1,0.2,0.3,0.1,0.2,0.3,0.1,0.2,0.3))

}
