% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelB4.R
\name{ModelB4}
\alias{ModelB4}
\title{Compute the maximum likelihood estimate of Model B4}
\usage{
ModelB4(x, x0, del, param)
}
\arguments{
\item{x}{Observation of the state variable at time t}

\item{x0}{Observation of the state variable at time t-1}

\item{del}{The time step between the current and previous observation}

\item{param}{The parameter 9-vector (a0,a1,a,b,L2,g,beta,f,r)}
}
\description{
Compute the maximum likelihood estimate of Model B4
}
\examples{
ModelB4(0.4,0.3,0.1,c(0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9))

}
