% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelB6.R
\name{ModelB6}
\alias{ModelB6}
\title{Compute the maximum likelihood estimate of Model B6}
\usage{
ModelB6(x, x0, del, param)
}
\arguments{
\item{x}{Observation of the state variable at time t}

\item{x0}{Observation of the state variable at time t-1}

\item{del}{The time step between the current and previous observation}

\item{param}{The parameter 5-vector (m, a, b, s, r):}
}
\description{
x1 is the log of the stock price, x2 the instantaneous variance.
}
\details{
dx1 <- (m - x2/2) dt + sqrt(x2) dW1
dx2 <- (a - b x2) dt + s sqrt(1-r^2) Sqrt(x2) dW1 + s r sqrt(x2) dW2
}
\examples{
ModelB6(0.4,0.3,0.1,c(0.1,0.2,0.3,0.3,-0.2))

}
