% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelU10.R
\name{ModelU10}
\alias{ModelU10}
\title{Compute the maximum likelihood estimate of Model U10}
\usage{
ModelU10(x, x0, del, param)
}
\arguments{
\item{x}{Observation of the state variable at time t}

\item{x0}{Observation of the state variable at time t-1}

\item{del}{The time step between the current and previous observation}

\item{param}{The parameter 8-vector (am1,a0,a1,a2,b0,b1,b2,b3)}
}
\description{
Compute the maximum likelihood estimate of Model U10
}
\examples{
ModelU10(0.4,0.3,0.1,c(0.1,0.3,0.2,0.1,0.2,1.3,0.1,0.2))

}
