% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelU11.R
\name{ModelU11}
\alias{ModelU11}
\title{Compute the maximum likelihood estimate of Model U11}
\usage{
ModelU11(x, x0, del, param)
}
\arguments{
\item{x}{Observation of the state variable at time t}

\item{x0}{Observation of the state variable at time t-1}

\item{del}{The time step between the current and previous observation}

\item{param}{The parameter 4-vector (a,b,f,d)}
}
\description{
Compute the maximum likelihood estimate of Model U11
}
\examples{
ModelU11(0.4,0.3,0.1,c(0.1,0.3,0.2,0.1,0.2,1.3))

}
