% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelU12.R
\name{ModelU12}
\alias{ModelU12}
\title{Compute the maximum likelihood estimate of Model U12}
\usage{
ModelU12(x, x0, del, param)
}
\arguments{
\item{x:}{Observation of the state variable at time t}

\item{x0:}{Observation of the state variable at time t-1}

\item{del:}{The time step between the current and previous observation}

\item{param:}{The parameter 3-vector (alpha,beta,gamma)}
}
\description{
Compute the maximum likelihood estimate of Model U12
}
\examples{
ModelU12(0.4,0.3,0.1,c(0.1,0.3,0.2))

}
