% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelU13.R
\name{ModelU13}
\alias{ModelU13}
\title{Compute the maximum likelihood estimate of Model U13}
\usage{
ModelU13(x, x0, del, param)
}
\arguments{
\item{x:}{Observation of the state variable at time t}

\item{x0:}{Observation of the state variable at time t-1}

\item{del:}{The time step between the current and previous observation}

\item{param:}{The parameter 7-vector (am1,a0,a1,a2,a3,sigma,rho)}
}
\description{
Compute the maximum likelihood estimate of Model U13
}
\examples{
ModelU13(0.4,0.3,0.1,c(0.1,0.3,0.2,0.1,0.2,1.3,0.5))

}
