% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelU2.R
\name{ModelU2}
\alias{ModelU2}
\title{Compute the maximum likelihood estimate of Model U2}
\usage{
ModelU2(x, x0, del, param, args = NULL)
}
\arguments{
\item{x}{Observation of the state variable at time t}

\item{x0}{Observation of the state variable at time t-1}

\item{del}{The time step between the current and previous observation}

\item{param}{The parameter 3-vector}

\item{args}{Not currently used}
}
\description{
Compute the maximum likelihood estimate of Model U2
}
\examples{
ModelU2(0.4,0.3,0.1,c(0.1,0.2,0.3))

}
