% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelU3.R
\name{ModelU3}
\alias{ModelU3}
\title{Compute the maximum likelihood estimate of Model U3}
\usage{
ModelU3(x, x0, del, param, args)
}
\arguments{
\item{x}{Observation of the state variable at time t}

\item{x0}{Observation of the state variable at time t-1}

\item{del}{The time step between the current and previous observation}

\item{param}{The parameter 4-vector (a,b,c,d)}

\item{args}{Not currently used}
}
\description{
linear drift, CEV diffusion (Univariate Model U3), dX = b(a - X) dt + c X^d dW
}
\examples{
ModelU3(0.4,0.3,0.1,c(0.1,0.2,0.3,0.1))

}
