% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelU4.R
\name{ModelU4}
\alias{ModelU4}
\title{Compute the maximum likelihood estimate of Model U4}
\usage{
ModelU4(x, x0, del, param)
}
\arguments{
\item{x}{Observation of the state variable at time t}

\item{x0}{Observation of the state variable at time t-1}

\item{del}{The time step between the current and previous observation}

\item{param}{The parameter 3-vector (kappa,alpha,sigma)}
}
\description{
CIR model: dX = kappa (alpha - X) dt + sigma sqrt(X) dW
}
\examples{
ModelU4(0.4,0.3,0.1,c(0.1,0.3,0.2))

}
