% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelU5.R
\name{ModelU5}
\alias{ModelU5}
\title{Compute the maximum likelihood estimate of Model U5}
\usage{
ModelU5(x, x0, del, param)
}
\arguments{
\item{x:}{Observation of the state variable at time t}

\item{x0:}{Observation of the state variable at time t-1}

\item{del:}{The time step between the current and previous observation}

\item{param:}{The parameter 6-vector (gamma,rho>=1,theta0,theta1,theta2,theta3)}
}
\description{
mu(x) <- theta0 + theta1*x + theta2*x^2 + theta3*x^3
s(x) <- gamma*x^rho
}
\examples{
ModelU5(0.4,0.3,0.1,c(0.1,1.3,0.2,0.3,0.2,0.1))

}
