% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelU6.R
\name{ModelU6}
\alias{ModelU6}
\title{Compute the maximum likelihood estimate of Model U6}
\usage{
ModelU6(x, x0, del, param)
}
\arguments{
\item{x:}{Observation of the state variable at time t}

\item{x0:}{Observation of the state variable at time t-1}

\item{del:}{The time step between the current and previous observation}

\item{param:}{The parameter 5-vector (a,b,c,d,f)}
}
\description{
mu[x_] = a + b*x + c*x^2 + d*x^3  
s[x_] = f
}
\examples{
ModelU6(0.4,0.3,0.1,c(0.1,0.3,0.2,0.3,0.1))

}
