% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelU7.R
\name{ModelU7}
\alias{ModelU7}
\title{Compute the maximum likelihood estimate of Model U7}
\usage{
ModelU7(x, x0, del, param)
}
\arguments{
\item{x}{Observation of the state variable at time t}

\item{x0}{Observation of the state variable at time t-1}

\item{del}{The time step between the current and previous observation}

\item{param}{The parameter 3-vector (kappa,alpha,eta)}
}
\description{
Vasicek (Ornstein-Uhlenbeck) Model
 mu = kappa (alpha-x);
 Sigma =eta;
}
\examples{
ModelU7(0.4,0.3,0.1,c(0.1,0.3,0.2))

}
