% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelU8.R
\name{ModelU8}
\alias{ModelU8}
\title{Compute the maximum likelihood estimate of Model U8}
\usage{
ModelU8(x, x0, del, param)
}
\arguments{
\item{x}{Observation of the state variable at time t}

\item{x0}{Observation of the state variable at time t-1}

\item{del}{The time step between the current and previous observation}

\item{param}{The parameter 6-vector (am1,a0,a1,a2,sigma,rho>=1)}
}
\description{
Compute the maximum likelihood estimate of Model U8
}
\examples{
ModelU8(0.4,0.3,0.1,c(0.1,0.3,0.2,0.1,0.2,1.3))

}
