% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Compute the summary information of the maximum likelihood estimate}
\usage{
summary(logdensity, x, del, param0, args = NULL)
}
\arguments{
\item{logdensity:}{the model log likelihood function}

\item{x:}{Time series of the observed state variables}

\item{del:}{The uniform time step between observations}

\item{param:}{The maximum likelihood estimated parameter}
}
\description{
Compute the summary information of the maximum likelihood estimate
}
\examples{
 summary(ModelU1,x,0.1,c(0.01,0.2))

}
